==> 01-ScrapingData.ipynb <==
# BeautifulSoup documentation
# https://www.crummy.com/software/BeautifulSoup/bs4/doc/

# Requests documentation
# https://requests.readthedocs.io/en/latest/

from bs4 import BeautifulSoup # for pulling html data
import requests # send http requests to url
from re import search

# Webscraping tutorial
# https://www.geeksforgeeks.org/implementing-web-scraping-python-beautiful-soup/

url = "https://www.nba.com/news/2022-nba-draft-order" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server 

# print(r.content) # gathers raw HTML content of webpage

soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure

# print(soup.prettify()) # prints nested structure of html content

table = soup.find('script', attrs = {'type':'application/ld+json'}) # all players and draft numbers and rounds found under script with attribute type = application/ld+json

# print(table.prettify()) # prints out the specified html content 

import json # extracting json from script tag
# https://morioh.com/p/4ed4ba1fe438

json_object = json.loads(table.contents[0]) # load the json into json_object

test = json_object['articleBody']

import re # regular expressions library

# https://stackoverflow.com/questions/32680030/match-text-between-two-strings-with-regular-expression
# matching text between two strings

m = re.findall(r'draft(.*?)(.*?)\(',test)
# m[0][1]

import pandas as pd # pandas for dataframe

# https://www.journaldev.com/23763/python-remove-spaces-from-string#:~:text=strip(),remove%20leading%20and%20trailing%20whitespaces.&text=If%20you%20want%20to%20remove,or%20rstrip()%20function%20instead.
# remove trailing and leading space

# store names in a list
testls = []
for name in m:
    testls.append(name[1].strip())

# making dataframe with the player names list
df = pd.DataFrame({'Player':testls})

# making columns of draft round and number picked

df['Round'] = 1 # setting round to 1 for now
df['DraftNumber'] = df.index + 1 # draft number is just the index plus one

# if draft number greater than 30, then round would be 2
df.loc[df['DraftNumber']> 30, 'Round'] = 2

# repeat for years 2021 and 2020

url = "https://www.nba.com/news/2021-nba-draft-results-picks-1-60" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find('div', attrs = {'class':'ArticleContent_article__NBhQ8'}) # all players and draft numbers and rounds found under div with attribute class = ArticleContent_article__NBhQ8

# https://scottlarsen.com/2020/08/19/Scraping-Web-Content-Between-Two-Tags-With-BeautifulSoup-And-Python.html
# for the get_text() function

list2021 = [] # empty list to store names

n = 0 # first 2 rows aren't names, thus have a counter 

for row in table.findAll('p'): # iterate through the table
    
    if n!=2: # if n isn't 2 then can't add row to list
        n = n + 1
        
    elif row.a == None: # last few rows are none and gave an error, came back to fix
        n = n
        
    else: # this step adds the name to the list 
        list2021.append(row.a.get_text())
    

# making dataframe with the player names list
df2021 = pd.DataFrame({'Player':list2021})

# making columns of draft round and number picked

df2021['Round'] = 1 # setting round to 1 for now
df2021['DraftNumber'] = df2021.index + 1 # draft number is just the index plus one
# if draft number greater than 30, then round would be 2
df2021.loc[df2021['DraftNumber']> 30, 'Round'] = 2

url = "https://www.nba.com/news/2020-nba-draft-results-picks-1-60" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find('div', attrs = {'class':'ArticleContent_article__NBhQ8'}) # all players and draft numbers and rounds found under div with attribute class = ArticleContent_article__NBhQ8

list2020 = [] # empty list to store names

n = 0 # first 2 rows aren't names, thus have a counter 

for row in table.findAll('a'): # iterate through the table
    
    if n!=2: # if n isn't 2 then can't add row to list
        n = n + 1
        
    elif row.a == None: # last few rows are none and gave an error, came back to fix
        n = n
        
    else: # this step adds the name to the list 
        list2020.append(row.get_text())

list2020 = []

for row in table.findAll('a'):
    # fixed bug where the html content had strings dealt or traded in between the a tag
    if row.get_text()[0].isupper(): # use isupper on first letter of string to see if it is a name
        list2020.append(row.get_text())



# making dataframe with the player names list
df2020 = pd.DataFrame({'Player':list2020})

# making columns of draft round and number picked

df2020['Round'] = 1 # setting round to 1 for now
df2020['DraftNumber'] = df2020.index + 1 # draft number is just the index plus one
# if draft number greater than 30, then round would be 2
df2020.loc[df2020['DraftNumber']> 30, 'Round'] = 2

# cleaned name that had \n after
df2020.loc[26,'Player'] = 'Udoka Azubuike'

url = "https://basketball.realgm.com/nba/draft/past-drafts" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find('tbody') # find pre league team

from re import search


i = 0 # counter for searching player name in dataframe
j = 0 # counter for every third row to get predraft team

# list to store predraft team
predraftteam = []

# iterate through all rows containing the tag a
for row in table.findAll('a'):
    
    # searches if first name is contained in player name in the table since some names were written differently between the two sources
    if search(df.Player[i].split()[0], row.get_text()): 
        j = 1 # makes counter 1 when player name is found
        
        # print(df.Player[i]) # to check to see if condition is met
    
    if j == 3: # every third row, add predraft team to list
        
        predraftteam.append(row.get_text()) # add predraft team to list
        i = i + 1 # iterate to the next player from first dataframe
    
    j = j + 1 # iterate counter

table1 = soup.find_all('tbody') # find pre league team

i = 30 # index 30 is the start of the second round
j = 0 # counter to get correct row to add pre draft team

# iterate through each row of second table with tag td
for row in table1[1].findAll('td'):
    
    # searches if first name is contained in player name in the table since some names were written differently between the two sources
    if search(df.Player[i].split()[1], row.get_text()) or search(df.Player[i].split()[0], row.get_text()):
        
        j = 1 # makes counter 1 when player name is found
    
    if j == 9: # every eighth row after name, add predraft team to list
        
        predraftteam.append(row.get_text()) # add predraft team to list
        
        # print(row.get_text()) # used print to see if for loop was working correctly
        
        i = i + 1 # iterate to the next player from first dataframe
    
    if i==58: # fixed error where index went out of range of dataframe
        i = 57
        
    j = j + 1 # iterate counter


predraftteam

df['PreTeam'] = predraftteam

df2021.head()

url = "https://basketball.realgm.com/nba/draft/past_drafts/2021" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find('tbody') # find pre league team

i = 0 # counter for searching player name in dataframe
j = 0 # counter for every third row to get predraft team

# list to store predraft team
predraftteam2021 = []

# iterate through all rows containing the tag a
for row in table.findAll('a'):
    
    # searches if last name is contained in player name in the table since some names were written differently between the two sources
    if search(df2021.Player[i].split()[1], row.get_text()): 
        j = 1 # makes counter 1 when player name is found
        
        # print(df2021.Player[i]) # to check to see if condition is met
    
    if j == 3: # every third row, add predraft team to list
        
        predraftteam2021.append(row.get_text()) # add predraft team to list
        i = i + 1 # iterate to the next player from first dataframe
    
    j = j + 1 # iterate counter

table1 = soup.find_all('tbody') # find pre league team

i = 30 # index 30 is the start of the second round
j = 0 # counter to get correct row to add pre draft team

# iterate through each row of second table with tag td
for row in table1[1].findAll('td'):
    
    # searches if first name is contained in player name in the table since some names were written differently between the two sources
    if search(df2021.Player[i].split()[1], row.get_text()) or search(df2021.Player[i].split()[0], row.get_text()):
        
        j = 1 # makes counter 1 when player name is found
    
    if j == 9: # every eighth row after name, add predraft team to list
        
        predraftteam2021.append(row.get_text()) # add predraft team to list
        
        # print(row.get_text()) # used print to see if for loop was working correctly
        
        i = i + 1 # iterate to the next player from first dataframe
    
    if i==60: # fixed error where index went out of range of dataframe
        i = 59
        
    j = j + 1 # iterate counter

df2021['PreTeam'] = predraftteam2021

url = "https://basketball.realgm.com/nba/draft/past_drafts/2020" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find('tbody') # find pre league team

i = 0 # counter for searching player name in dataframe
j = 0 # counter for every third row to get predraft team

# list to store predraft team
predraftteam2020 = []

# iterate through all rows containing the tag a
for row in table.findAll('a'):
    
    # searches if first/last name is contained in player name in the table since some names were written differently between the two sources
    if search(df2020.Player[i].split()[1], row.get_text()) or search(df2020.Player[i].split()[0], row.get_text()): 
        j = 1 # makes counter 1 when player name is found
        
        # print(df2020.Player[i]) # to check to see if condition is met
    
    if j == 3: # every third row, add predraft team to list
        
        predraftteam2020.append(row.get_text()) # add predraft team to list
        i = i + 1 # iterate to the next player from first dataframe
    
    j = j + 1 # iterate counter


table1 = soup.find_all('tbody') # find pre league team

i = 30 # index 30 is the start of the second round
j = 0 # counter to get correct row to add pre draft team

# iterate through each row of second table with tag td
for row in table1[1].findAll('td'):
    #print(j) # finding bug where loop didn't capture the last 3 predraft teams
    #print(row.get_text()) # finding bug where loop didn't capture the last 3 predraft teams
    # searches if first/last name is contained in player name in the table since some names were written differently between the two sources
    if (search(df2020.Player[i].split()[1], row.get_text()) or search(df2020.Player[i].split()[0], row.get_text())) and row.get_text()!="DePaul": 
        # School name DePaul was matched as the name Paul Reed Jr., added an AND condition to fix bug
        
        j = 1 # makes counter 1 when player name is found
    
    if j == 9: # every eighth row after name, add predraft team to list
        
        predraftteam2020.append(row.get_text()) # add predraft team to list
        
        #print(row.get_text()) # used print to see if for loop was working correctly
        i = i + 1 # iterate to the next player from first dataframe
    
    if i==60: # fixed error where index went out of range of dataframe
        i = 59
        
    j = j + 1 # iterate counter

df2020['PreTeam'] = predraftteam2020

df2020.head()

url = "https://basketball.realgm.com/nba/draft/past_drafts/2020" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure

# inspiration from https://stackoverflow.com/questions/5815747/beautifulsoup-getting-href

table1 = soup.find_all('tbody') # find pre league team

# empty list to store url links to player summary
urllist2020 = []

# outside loop to access the first and second round players
for i in range(0,2):
    
    # finding all a tags with reference to links to find link for player summary
    for row in table1[i].find_all('a', href = True):
        
        # search for links that contain /player to find player stats
        if search("/player",row['href']):
            
            # add the main website and the slash after it to list
            url = "https://basketball.realgm.com"
            urllist2020.append(url+row['href'])

df2020['urlsummary'] = urllist2020

url = "https://basketball.realgm.com/nba/draft/past_drafts/2021" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table1 = soup.find_all('tbody') # find pre league team

# empty list to store url links to player summary
urllist2021 = []

# outside loop to access the first and second round players
for i in range(0,2):
    
    # finding all a tags with reference to links to find link for player summary
    for row in table1[i].find_all('a', href = True):
        
        # search for links that contain /player to find player stats
        if search("/player",row['href']):
            
            # add the main website and the slash after it to list
            url = "https://basketball.realgm.com"
            urllist2021.append(url+row['href'])

df2021['urlsummary'] = urllist2021

url = "https://basketball.realgm.com/nba/draft/past_drafts/2022" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table1 = soup.find_all('tbody') # find pre league team

# empty list to store url links to player summary
urllist2022 = []

# outside loop to access the first and second round players
for i in range(0,2):
    
    # finding all a tags with reference to links to find link for player summary
    for row in table1[i].find_all('a', href = True):
        
        # search for links that contain /player to find player stats
        if search("/player",row['href']):
            
            # add the main website and the slash after it to list
            url = "https://basketball.realgm.com"
            urllist2022.append(url+row['href'])

df['urlsummary'] = urllist2022

# Setting preteam type to college as majority of draft players played in NCAA
df['PreType'] = "College"

# Going through all the teams in the dataframe
for team in df['PreTeam']:
    
    # Overseas team has brackets after the international team indicating country
    if search("\(",team):
        df.loc[df['PreTeam'] == team,'PreType'] = "Overseas"
    
    # G League Players that got drafted came from NBA G League Ignite
    elif search("NBA",team):
        df.loc[df['PreTeam'] == team, 'PreType'] = "G League"

# Setting preteam type to college as majority of draft players played in NCAA
df2021['PreType'] = "College"

# Going through all the teams in the dataframe
for team in df2021['PreTeam']:
    
    # Overseas team has brackets after the international team indicating country
    if search("\(",team):
        df2021.loc[df2021['PreTeam'] == team,'PreType'] = "Overseas"
    
    # G League Players that got drafted came from NBA G League Ignite
    elif search("NBA",team):
        df2021.loc[df2021['PreTeam'] == team, 'PreType'] = "G League"

# Setting preteam type to college as majority of draft players played in NCAA
df2020['PreType'] = "College"

# Going through all the teams in the dataframe
for team in df2020['PreTeam']:
    
    # Overseas team has brackets after the international team indicating country
    if search("\(",team):
        df2020.loc[df2020['PreTeam'] == team,'PreType'] = "Overseas"
    
    # G League Players that got drafted came from NBA G League Ignite
    elif search("NBA",team):
        df2021.loc[df2020['PreTeam'] == team, 'PreType'] = "G League"

# for url in df['urlsummary']:
#     r = requests.get(url) # send http request to url and saves response from server
#     soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
#     table1 = soup.find_all('tbody') # find pre league team



########### Find common pattern to get stats

url = 'https://basketball.realgm.com/player/Ousmane-Dieng/Summary/148828'
r = requests.get(url)
soup = BeautifulSoup(r.content, 'html.parser')
table1 = soup.find_all('tbody')

# inspired by https://stackoverflow.com/questions/67938081/use-beautifulsoup-to-find-a-table-after-a-header

# checking before running loop through all players

checker = 0
numberofstats = 1
    
for i in range(0,len(table1)):
    for row in table1[i].findAll('td'):
        if row.get_text() == '2021-22':
            checker = 1
        elif checker == 1:
            if any(c.isalpha() for c in row.get_text()) == False and numberofstats <= 21:
                print(float(row.get_text()))
                numberofstats = numberofstats + 1

# checker = 0
# numberofstats = 1

# for row in table1.findAll('td'):
#     if row.get_text() == '2021-22':
#         checker = 1
#     elif checker == 1:
#         if any(c.isalpha() for c in row.get_text()) == False and numberofstats <= 21:
#             print(float(row.get_text()))
#             numberofstats = numberofstats + 1


# list to store all draft players stats
totalls = []

# looping through all draft players summary url
for url in df['urlsummary']:
    
    # send http request to url and saves response from server
    r = requests.get(url)
    
    # represents document as nested data structure
    soup = BeautifulSoup(r.content, 'html.parser')
    
    # conditions to check whether player came from college, overseas or g league
    # then finding the table containing stats per game under the respective h2 tag
    if (df.loc[df['urlsummary']==url,'PreType'] == 'College').bool():
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "NCAA Season Stats - Per Game" in tag.text).find_next("table")
    
    elif (df.loc[df['urlsummary']==url,'PreType'] == 'Overseas').bool():
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "International Regular Season Stats - Per Game" in tag.text).find_next("table")
    
    else:
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "G League Full Season Stats - Per Game" in tag.text).find_next("table")
    
    # list to store individual player stats in
    playerls = []
    
    # checker for finding the right year which is 2021-22, the year previous of the draft
    checker = 0
    
    # checker for having the correct number of stats in the table
    numberofstats = 1
    
    # iterate through the table containing the tag td
    for row in tableh2tag.findAll('td'):
        
        # find the right year and change the checker to 1
        if row.get_text() == '2021-22' or row.get_text() == '2021-22 *':
            checker = 1
        
        # if the checker is 1 then all the rows below are the stats in that year
        elif checker == 1:
            
            # condition for letters as we only want numbers, the beginning rows of the table 
            # contained team name and league, second condition makes sure that it retrieve 
            # the correct number of stats which is 21
            if any(c.isalpha() for c in row.get_text()) == False and numberofstats <= 21:
                
                # saving the stat in a variable
                placeholder = row.get_text()
                
                # there was an issue where a player didn't play in college but still got drafted (Shaedon Sharpe)
                # thus his stats were a dash, so change all the dashes to 0 to represent empty or nothing
                if row.get_text() == '-':
                    placeholder = 0
                    
                # changing the type to float as the text came out as strings
                playerls.append(float(placeholder))
                
                # add 1 to this checker after adding one stat to the list
                numberofstats = numberofstats + 1
    # add list of stats of one player to the total list
    totalls.append(playerls)
    

# Shaedon Sharpe didn't play a single game at Kentucky but still declared for the draft
# Still was considered as high lottery player

# Could study highschool stats to see if highschool translate to NBA

# totalls[6] = ['0']*21

import numpy as np

# 1. gather header names for stats

url = 'https://basketball.realgm.com/player/Paolo-Banchero/Summary/134150'
r = requests.get(url)
soup = BeautifulSoup(r.content, 'html.parser')
table1 = soup.find_all('thead')

statheaderls = []

for row in table1[0].findAll('th'):
    if len(row.get_text()) < 4:
        statheaderls.append(row.get_text())

statheaderls





# 2. add column names to df

# testdf = df.copy()

# 3. add array of stats to df 
df[statheaderls] = np.array(totalls)

# to check for empty lists 

# ran into an issue with some overseas players that played in different leagues 
# issue was the year was 2021-22 * and not 2021-22 so added the condition to the for loop above

checkls = []
for i in range(0,len(totalls)):
    if len(totalls[i]) == 0:
        checkls.append(i)
checkls

# issue with Loyola(MD) with being college but was seen as overseas because
# the pattern with overseas on this site is that usually overseas team have
# team(country) so the brackets in my earlier code found this college to be overseas

df2021.loc[df2021['Player']=="Santi Aldama",'PreType'] = "College"

# list to store all draft players stats
totalls = []

# looping through all draft players summary url
for url in df2021['urlsummary']:
    
    # send http request to url and saves response from server
    r = requests.get(url)
    
    # represents document as nested data structure
    soup = BeautifulSoup(r.content, 'html.parser')
    
    #print(url) # checking error, issue where table under header is different this year than 2022
    
    # conditions to check whether player came from college, overseas or g league
    # then finding the table containing stats per game under the respective h2 tag
    if (df2021.loc[df2021['urlsummary']==url,'PreType'] == 'College').bool():
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "NCAA Season Stats - Per Game" in tag.text).find_next("table")
    
    elif (df2021.loc[df2021['urlsummary']==url,'PreType'] == 'Overseas').bool():
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "International Regular Season Stats - Per Game" in tag.text).find_next("table")
    
    else:
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "G League Regular Season Stats - Per Game" in tag.text).find_next("table")
    
    # list to store individual player stats in
    playerls = []
    
    # checker for finding the right year which is 2020-21, the year previous of the draft
    checker = 0
    
    # checker for having the correct number of stats in the table
    numberofstats = 1
    
    ######################
    # issue with index 29#
    ######################
    # loyola(md) was classified as overseas because of the brackets
    # fixed in code in the chunk above
    
    # iterate through the table containing the tag td
    for row in tableh2tag.findAll('td'):
        # find the right year and change the checker to 1
        if row.get_text() == '2020-21' or row.get_text() == '2020-21 *':
            checker = 1
        
        # if the checker is 1 then all the rows below are the stats in that year
        elif checker == 1:
            
            
            # condition for letters as we only want numbers, the beginning rows of the table 
            # contained team name and league, second condition makes sure that it retrieve 
            # the correct number of stats which is 21
            if any(c.isalpha() for c in row.get_text()) == False and numberofstats <= 21:
                
                # saving the stat in a variable
                placeholder = row.get_text()
                
                # there was an issue where a player didn't play in college but still got drafted (Shaedon Sharpe)
                # thus his stats were a dash, so change all the dashes to 0 to represent empty or nothing
                if row.get_text() == '-':
                    placeholder = 0
                    
                # changing the type to float as the text came out as strings
                playerls.append(float(placeholder))
                
                # add 1 to this checker after adding one stat to the list
                numberofstats = numberofstats + 1
    # add list of stats of one player to the total list
    totalls.append(playerls)

# adding array of stats to the dataframe with the respective stat header names
df2021[statheaderls] = np.array(totalls)

# issue with IMG Academy(Florida) with being prep school but was seen as overseas because
# the pattern with overseas on this site is that usually overseas team have
# team(country) so the brackets in my earlier code found this prep school to be overseas

df2020.loc[df2020['Player']=="Kenyon Martin Jr",'PreType'] = "Prep School"

# list to store all draft players stats
totalls = []

# looping through all draft players summary url
for url in df2020['urlsummary']:
    
    # send http request to url and saves response from server
    r = requests.get(url)
    
    # represents document as nested data structure
    soup = BeautifulSoup(r.content, 'html.parser')
    
    # print(url) # checking error, issue where table under header is different this year than 2022
    
    # conditions to check whether player came from college, overseas or g league
    # then finding the table containing stats per game under the respective h2 tag
    if (df2020.loc[df2020['urlsummary']==url,'PreType'] == 'College').bool():
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "NCAA Season Stats - Per Game" in tag.text).find_next("table")
    
    elif (df2020.loc[df2020['urlsummary']==url,'PreType'] == 'Overseas').bool():
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "International Regular Season Stats - Per Game" in tag.text).find_next("table")
    
    else:
        tableh2tag = soup.find(lambda tag: tag.name == "h2" and "G League Regular Season Stats - Per Game" in tag.text).find_next("table")
    
    
    # list to store individual player stats in
    playerls = []
    
    # checker for finding the right year which is 2019-20, the year previous of the draft
    checker = 0
    
    # checker for having the correct number of stats in the table
    numberofstats = 1
        
    
    # iterate through the table containing the tag td
    for row in tableh2tag.findAll('td'):
        # find the right year and change the checker to 1
        if row.get_text() == '2019-20' or row.get_text() == '2019-20 *':
            checker = 1
        
        # if the checker is 1 then all the rows below are the stats in that year
        elif checker == 1:
            
            
            # condition for letters as we only want numbers, the beginning rows of the table 
            # contained team name and league, second condition makes sure that it retrieve 
            # the correct number of stats which is 21
            if any(c.isalpha() for c in row.get_text()) == False and numberofstats <= 21:
                
                # saving the stat in a variable
                placeholder = row.get_text()
                
                # there was an issue where a player didn't play in college but still got drafted (Shaedon Sharpe)
                # thus his stats were a dash, so change all the dashes to 0 to represent empty or nothing
                if row.get_text() == '-':
                    placeholder = 0
                    
                # changing the type to float as the text came out as strings
                playerls.append(float(placeholder))
                
                # add 1 to this checker after adding one stat to the list
                numberofstats = numberofstats + 1
    # add list of stats of one player to the total list
    totalls.append(playerls)

# adding array of stats to the dataframe with the respective stat header names
# df2020[statheaderls] = np.array(totalls)

# KJ Martin played in IMG academy (prep school) so no stats were available

totalls[51] = ['0']*21

df2020[statheaderls] = np.array(totalls)


df2020

# df.to_csv('data/df2022.csv')

# df2021.to_csv('data/df2021.csv')

# df2020.to_csv('data/df2020.csv')

# import pandas as pd

# df2022 = pd.read_csv("data/df2022.csv")
# df2021 = pd.read_csv("data/df2021.csv")
# df2020 = pd.read_csv("data/df2020.csv")

df2022 = df
df2022.head()

url = "https://basketball.realgm.com/nba/draft/past_drafts" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find_all('tbody') # find pre league team

i = 0 # counter for searching player name in dataframe
j = 0 # counter for every fourth row to get position

# list to store position
position = []

# iterate through all rows containing the tag td
for k in range(0,2):
    for row in table[k].findAll('td'):
            # searches if first name is contained in player name in the table since some names were written differently between the two sources
        if search(df2022.Player[i].split()[1], row.get_text()) or search(df2022.Player[i].split()[0], row.get_text()):
        
            j = 1 # makes counter 1 when player name is found
    
        if j == 4: # every eighth row after name, add position to list
        
            position.append(row.get_text()) # add position to list
        
            #print(row.get_text()) # used print to see if for loop was working correctly
        
            i = i + 1 # iterate to the next player from first dataframe
    
        if i==58: # fixed error where index went out of range of dataframe
            i = 57
        
        j = j + 1 # iterate counter

df2022['Position'] = position

url = "https://basketball.realgm.com/nba/draft/past_drafts/2021" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find_all('tbody') # find pre league team


i = 0 # counter for searching player name in dataframe
j = 0 # counter for every fourth row to get position

# list to store position
position = []

for k in range(0,2):
    # iterate through each row of second table with tag td
    for row in table[k].findAll('td'):

        # searches if first name is contained in player name in the table since some names were written differently between the two sources
        if search(df2021.Player[i].split()[1], row.get_text()) or search(df2021.Player[i].split()[0], row.get_text()):
            if i!=59:
                i = i + 1 # iterate to the next player from first dataframe
            j = 1 # makes counter 1 when player name is found
            #print(row.get_text())

        if j == 4: # every fourth row after name, add position to list

            position.append(row.get_text()) # add position to list

            #print(row.get_text()) # used print to see if for loop was working correctly

        #if i==60: # fixed error where index went out of range of dataframe
            #i = 59

        j = j + 1 # iterate counter
        
df2021['Position'] = position

url = "https://basketball.realgm.com/nba/draft/past_drafts/2020" # url for the players that got drafted
r = requests.get(url) # send http request to url and saves response from server
soup = BeautifulSoup(r.content, 'html.parser') # represents document as nested data structure
table = soup.find_all('tbody') # find pre league team


i = 0 # counter for searching player name in dataframe
j = 0 # counter for every fourth row to get position

# list to store position
position = []

for k in range(0,2):
    # iterate through each row of second table with tag td
    for row in table[k].findAll('td'):

        # searches if first name is contained in player name in the table since some names were written differently between the two sources
        if search(df2020.Player[i].split()[1], row.get_text()) or search(df2020.Player[i].split()[0], row.get_text()) and row.get_text()!="DePaul":
            if i!=59:
                i = i + 1 # iterate to the next player from first dataframe
            j = 1 # makes counter 1 when player name is found
            #print(row.get_text())

        if j == 4: # every fourth row after name, add position to list

            position.append(row.get_text()) # add position to list

            #print(row.get_text()) # used print to see if for loop was working correctly

        #if i==60: # fixed error where index went out of range of dataframe
            #i = 59

        j = j + 1 # iterate counter
        
df2020['Position'] = position

# df2022.to_csv('data/df2022.csv', index = False)
# df2021.to_csv('data/df2021.csv', index = False)
# df2020.to_csv('data/df2020.csv', index = False)

# testdf = pd.read_csv('data/df2022.csv')
# testdf
